* Still a work in progress!
* Added some functions for debugging TowerTree. DebugDrawRelationship() draws a series of lines and points, showing the
hierarchy of the tree. DebugLogHierarchy() logs the hierarchy node-by-node (so any infinite recursions will leave a wake of
logs behind them). Needless to say these functions are really slow and should never be used beyond testing purposes. However
at this point there's no way to disable them, this'll be added shortly.

* Added ReParentNode() to TowerTree. It can recursively travel up a tree and reassign each nodes' parent and children correctly.

* When removing a block, a node's children will now correctly work their way down the tree, checking for any potential adjacent
parent blocks, and will reconfigure themselves to conform with the new parent if they find one.

* Removing a block will correctly orphan any blocks that can't find a path to root. What to do with orphans is still not
implemented.

- Note: I just noticed pretty much every function in TowerTree works recursively. Looking up the Wikipedia article it seems
recursion is basically the easiest way to handle tree traversal. I had no idea about this and I never used recursion outside
of schoolwork so I'm kinda surprised I accidentally did it this way!

* The Flash part of the HUD no longer handles mouse clicks. It's now done through the regular input system which passes the
information along to the HUD. This has the benefit of probably being faster, tied to the game instead of the SWF, and supporting
more than the left mouse button.

- Note: I found out that, starting with the March UDK version, Flash HUDs will work with iPod/iPhone/iPads just fine. Good to
know because I was kind of considering starting work on a Canvas HUD!