* Still work in progress.
+ I mentioned last commit how I was impressed I implemented the tree recursively. Found out today UnrealScript is very very
slow when it comes to recursion, iteration is the way to go. So each function will have to be made to use iteration at some
point in the future! To put some perspective on how much slower recursion probably is: Unreal Engine allows 1,000,000
iterations of a loop before it considers it infinite and crashes, whereas it permits a recursive function to call itself only
250 times, 4000 times less than the max iteration.

* Added two configurable bools: bDebugDrawHierarchy and bDebugLogHierarchy to control debugging of the tree. Fully implemented.

* Block dropping seems to mostly work. The function for getting the time to check for new parents could use some fine-tuning,
but its good enough for now.

* TowerBlocks now have two states: Stable and Unstable. Stable is normal behavior, Unstable occurs when it has no path to root,
and thus it drops down due to gravity.

* Left/RightClickDown/Up() functions removed and replaced with a ClickDown() and ClickUp() function that pass along a numeric
ID for the mouse button clicked/released.

+ Did some script speed tests. Needless to say there are some big frame hitches when you have, say, a ~50 node branch and remove
the support under it. A console command, DebugCreateStressTower() will be used to generate a huge, stressful tower, and tests
will be done between the recursive implementation and iterative implementation.