* AddBlock() from TowerGame and up now returns the added TowerBlock, to help facilitate iteration vs recursion.
- Did some experiments with SuggestTossVelocity() for projectiles. Not really understanding what makes it fail and what makes it
work. It's clearly related to distance and velocity, but I don't get why it stops working when it stops working.

* Added a NodeCount int to TowerTree to keep track of the number of nodes. Mostly used in saving/loading.

* DebugCreateStressTower() removed because I'm dumb. Why make a function to randomly generate a huge tower, when I can just build
the tower, use a save system to save it, and then load it whenever I want? As such this build mostly focused on saving and loading
games (which is all done in a C++ DLL I don't commit (which I probably should) so it looks like I did pretty much nothing!). So
far it supports saving and loading: The version of the save file format (anticipated to be used to allow the loading of old
save formats), a bool (technically unsigned char) of whether this is a file of just tower data or also game data (probably
anticipated to allow multiplayer clients to save their towers since they won't have access to all game data), the number of
blocks in their tower (for accurately reading all block data), length of the name of the tower (for accurately reading the
actual tower name, which has a variable amount of characters), and as many wchar_ts as specified by the previous value.