* Did work on build menu in HUD. It's currently capable of rendering the in-game preview scene to a texture on the HUD. Now just need to get the list functioning and interacting with it properly.
* Mods have been moved from TowerGame to TowerPlayerReplicationInfo. TowerGame isn't available to clients, so they wouldn't have
been able to load mods in multiplayer. TowerGame still checks for mods (and will set the allowed mods for the server) through
means of locally spawning its own TowerPlayerReplicationInfo and destroying it once it gets the mod information.

* TowerMapInfo is being expanded to hold actors used for HUD preview rendering, so that they can be turned on and off as needed.

+ Things like BlockInfo and DisplayName and anything else at least tangently related to modding I added isn't set in stone.
Thinking about mod replication is a nightmare, not a technical one though. I could have class per block, making it easy to
replicate and organize, but it kind of defeats the purpose of classes (and especially components) to have children that differ
by maybe two or three lines of code. I could use the BlockInfo struct, so the vast majority of the blocks would be the same class
with just the modified settings in the BlockInfo struct. But then that could become a huge pain for organizing and keeping track
of all these blocks that could be their own class or that could be BlockInfos based on other classes... I'll figure it all out
soon, I'm kinda leaning towards BlockInfo just to avoid a million-class hierarchy of blocks.

- Also HUDMode in TowerHUD is now deprecated since there isn't really a Add/Remove mode anymore.