* Shuffling around mod stuff that will probably be shuffled around again. This time everything went from TowerGame to TowerPlayerReplicationInfo for the most part.
* TowerSaveSystem has all references to the DLL removed, it solely uses native UnrealScript code now.

* Worked on verifying clients have the mods before letting them join, and also replicated mod lists so clients know what blocks
to add to the HUD list and such.

* Worked on populating the list in the HUD. Works great on server, because it correctly follows an execution order of
TowerGame->TowerPlayeReplicationInfo->TowerHUD. Clients however, execute someway like TowerHUD->TowerPlayerReplicationInfo.
Thinking about it now that's probably because TowerPlayerReplicationInfo has to be created on the server and then replicated to
the client, but TowerHUD is all local so it executes first. That order is very specific though. For the server, TPRI exists
before TowerHUD, and for the client, TowerHUD exists before TPRI. So now I have to figure out how to get this to work for both
cases. Yay.

- Note for the future, BlockInfo will be the primary way to handle blocks, not making a new class for EVERY single new block.
Modules still undecided.