* Successfully upgraded to February 2011 UDK.
* UnrealScript now gets notifed when a name in the PlaceablesList is clicked on, which includes the index for that list slot.
Assuming proper list format is kept (and it's not even finalized yet), it's easy to compare the given index to the arrays to
determine what block (and module in the future) is selected.

* Because of the previous point, the preview block now changes based on what you select! The server and client seem to just
modify their own local copy, as they can have two different previews going on simultaneously, presumebly because I'm not
modifying ReplicatedMesh/ReplicatedMaterial. It looks pretty awesome if I may say so myself.

- Temporarily added a ModuleInfo to be consistent at this point. No idea if it'll change or what.

- Next comes implementing BlockInfo into everything (namely AddBlock, etc), and then working on placing modules!