* Converted all the annoying 'XBlock, YBlock, ZBlock' things sprinked through every AddBlock()-related code into a single vector called GridLocation. Much nicer.
* NextNodes has been removed. The whole PreviousNode/NextNodes thing is completely replaced with Base.

- Realized something about BlockInfo. The ability to base a new block off of a base class, but tweaking its parameters. I forgot
that already existed. It's called an archetype, you can make them in the editor, save them to packages, and reference them from
the code. I'll have to look into this real shortly.

* CollidingActors()'s radius was shrunk so it no longer picks up diagonal blocks when it iterates.

- Block falling still isn't perfect yet, no doubt thanks to revamping the whole system. It's getting there though.

* The way the mouse interacts with the HUD is completely changed now. There's a new class called GFxScrollingList, which is used
to represent ScrollingLists in Flash (which is what the BuildMenu PlaceablesList is). This class receives mouse events from
UnrealScript (things like mouse roll on/off, click/release, etc), and uses them to fake mouse events to its ItemRenderers in
Flash. It seems to work flawlessly, although not all mouse events are generated yet. Thanks to this though, the BuildMenu works
completely and you can click things in the PlaceablesList just like when the regular mouse was used.