* Projectiles pass through units of the same team.
* Hitscan weapons may pass through units of the same team, untested.

- Did some stress testing and a ton of GamplayProfiler and StatsViewer stuff.

* Now when the Begin state code of TowerEnemyController, a check is done to see if its pawn is within some arbitrary distance of the target. If so, it goes to
an idle state. Otherwise NavigationHandle.ActorReachable() is spammed thousands of times a second by all these infantry loitering next to the target, which is
apparently really expensive at such quantities.

* ScriptGetTeamNum() is the primary way of getting the TeamIndex of some unit. GetTeamNum() tries to go through PRI, which units obviously don't have.

* TowerFaction now extends TeamInfo, and TowerFactionHuman was added.

* Formations actually consume a cost!

* TowerGame kinda switched around to treat all players and AIs as having a faction and properly setting up TeamIndexes and teams in the GRI.

- As for testing, the biggest issue to the number of units alive at once is pawn ticking. 72.433 ms of the GameThread is spent on 301 calls to "Pawn Physics Time"
which is a part of "Pre AW (Asynchronous Work?) Actor Tick". Disabling tick on pawn and controller leaves the GameThread doing basically nothing extra, but
the rendering thread gets bogged down by 11.745 ms of "Dynamic Primitive drawing" and 13 ms of "Occlusion Result". With ticking, 90 pawns and controllers was
mostly fine, 300 was pushing maybe 7 FPS. Without ticking, things were definitely running better for the most part, and I think with 300 it was going ~30 FPS.
Disabling ticking for pawns is mostly impossible they can't do any sort of pathfinding or animation without it, so the best bet is simply to limit the number
of living units at once.

+ Short-term to do list!:

- Try out Actor-based modules!