* TowerAIObjectives will be a linked list and has a NextObjective variable to anticipate that. Its mesh is no longer drawn.
- Removed more dumb logging and ScriptTraces.

* Added debug 'spectating' for TowerTargetables and TowerFactionAIs. Basically it switches the camera view to their's, and adds them to PostRenderedActors. No
clue on how FactionAIs can actually adjust the camera though. So far for infantry it just shows basic info, but I would like to expand it to show leader, path,
etc.

- Removed TroopBudget from TowerFactionAI, it'll just use Budget like any other TowerFaction.

* Organized TowerGame into states (CoolDown and RoundInProgress) so there aren't a bunch of functions like StartCoolDown(), StartRound(), etc that made it kinda
confusing on when to call what. Also removed TowerGameReplicationInfo's round replication and HUD updating, and just made them into server->client RPC calls in
TowerPlayerController.

* Rounds actually mostly flow now. When every FactionAI's troops die and it doesn't have enough points to spawn more, it tells TowerGame. If every faction
(minus humans') tells TowerGame that, it goes to CoolDown, and after the CoolDown timer expires, the next round starts and it keeps looping like such. Only
issue is that if the AI never had enough Budget to spawn things, it never tells TowerGame it's done for the round.

- Added some simple precompile checks to start to anticipate demo and non-debug releases.

- Added some None checks for PostRenderedActors and TowerModuleGun's shooting.

* Added an exec function to kill every Targetable. We'll need one to actually skip a round at some point.

+ Todo!:

- Derive TowerBlock from Actor instead of DynamicSMActor, give it a MeshComponent variable as blocks can have either static or skeletal meshes.

- Properly implement A*!

- Fix TowerFactionAI bug mentioned above!

- Only allow saving in CoolDown!

- Take the A* paths and have the AI follow through with it! Doesn't matter if paths are dumb as long as it works!