* After a bunch of experimenting, block replication seems to work?! Clients can add/remove blocks, and see falling/reattaching blocks just fine!
* bFallingParent and bFallingChild have been removed from TowerBlockStructural. Clients now switch states according do their own simulations.

* ReplicatedBase is now the core of block simulation. Clients have no access to nearly all Tower functions, so the server handles determining the block bases
and whatnot. When the server does an important SetBase(), it also sets ReplicatedBase to that value. The client is replicated that value and sets its copy's
base accordingly. If the new base is now None, the client's block goes into the UnstableParent state and simulates the falling, very similar to what the
server is doing. The client is able to detect when the block hits the "ground" and change states properly, and transitions to the Stable state when
ReplicatedBase becomes something other than None. The native BaseChange() event is used by states to catch a base change, although its ignored on the server.

* TowerBlock::GetLocation()/GetGridLocation(), TowerBlockStructural::TimeToDrop()/IsTouchingGround(),
and TowerBlockStructural::BeginState()/EndState()/DroppedSpace() are now simulated for clients!

* bReplicateMovement for blocks is now false and must never be changed back! For any reason!

* TowerPlayerController now checks to see if essential actors have replicated yet (so far GRI and player's Tower) before requesting the mod list.

* Fixed bug where blocks didn't recognize root blocks owned by other players as being a valid root.

+ ToDo (Networking Edition!):

- Fix the bug with clients having to recalculate block locations 2-3 times to get proper results.

- Make sure basic TowerModule replication works.

- Replicate TowerModules shooting stuff.

- Replicate enemy Pawns and their shooting. Try to actually simulate and not devour all available bandwidth?

- Make sure dedicated servers work.

- Move the clients StopMovie() call to a logical position, preferably one that covers up all the waiting around for replication.