* TowerRangeComponent now holds the socket name and center point to be used by any derived class (similar variables removed from RangeCylinderComponent). Similar thing done for TowerAttackComponent.
* Fixed bug in Tower::FindNewParent() that allowed blocks to set a module as their base.

* Added TowerAttackComponent support for attaching particle systems to StaticMeshes.

* Fixed minor accessed none error for servers when closing the game.

* Updating client block locations works flawlessly in a single call! The problem is related to RelativeLocation, which I've had plenty of troubles with in
the past. Instead of trying to work within the bounds of RelativeLocation though, I bruteforced it since this function should only ever be called once a
game. Create an array of all the blocks, an array of all those blocks' base, set all bases to None, adjust location, set all block bases back to normal. Easy!
This is done in Tower::ReCalculateAllBlockLocations().

* Did some work loading levels with dedicated servers, but there's a few questions to answer first. Are the level blocks all loaded when the server starts?
Then who gets ownership of the blocks? A special server Tower instance? Does it transfer ownership to the first player that joins, or keep it? If everyone
leaves do you keep running that game? Reload it? Pause it?

+ ToDo (Networking Edition!):

- Not networking related, but I think it's time to lose the Tower prefix. Tower is already the package name, we don't need class'Tower.TowerBlock' and so forth.

- Replicate TowerModules shooting stuff.

- Replicate enemy Pawns and their shooting. Try to actually simulate and not devour all available bandwidth?

- Add level loading functionality for dedicated servers.

- Move the clients StopMovie() call to a logical position, preferably one that covers up all the waiting around for replication.