* Tackled converting TowerBlockStructural::IsTouchingGround from recursion to iteration, succeeded. Tested speeds and found that in the worst case (0 children), iteration was about 0.4x as fast as recursion. Approaching the best case though (310 children), iteration was about 1.5x faster. Of course iteration has the extra advantage that it doesn't crash the game when you reach a few-hundred block chain. Is not in use yet.
* UpdateGridLocation() and UpdateLocation() added to replace the awful SetGridLocation(). Neither in use yet.

* Added testing function to the CheatManager for benchmarking recursive and iterative ::IsTouchingGround().
Unfortunately the recursive one crashes the game so often that it's difficult to do more than 100 test iterations
at once when you have a lot of blocks.

* Player pawns probably going away. For whatever reason there's a lot of mismatch between the client and server when
it comes to moving player pawns, which is likely somewhat responsible for the jerkiness. For that reason it now uses
a spectating-like state of the PlayerController to represent players. I believe clients are authoratative concerning
their locations and don't do any mismatch checks, which is fine in a game where everyone's on the same team and
doesn't directly interact with the other. My only concern is I originally made the player pawns to fix replication
issues where it never saw enemies as relevant. Doesn't seem to be an issue now though?

* Added a bUsePlayerPawns config variable just so I don't have to keep cutting out/pasting code to jump between pawns
and no pawns when testing.

* Resolved bug in IsTouchingGround where children of whatever touched didn't have their GridLocation corrected. The
fix is to not just shove SetGridLocation() wherever needed, but to update it intelligently.

+ ToDo (Networking Edition!):

- Investigate weird jerkiness on clients when blocks are removed from underneath other blocks?

- Investigate other block anomalies. Expand UberTest if needed.

- Not networking related, but I think it's time to lose the Tower prefix.

- Replicate TowerModules shooting stuff.