* Upgraded to February 2012 UDK.
* Fixed non-deferred searching code. Since doing non-deferred searches assumably means you want the results
NOW, such searches call the completion delegate before StartGeneratePath() returns, as opposed to next tick.

* Fixed bug where paths that didn't finish at the given Finish point caused a ton of errors.

* Fixed potential bug where, by not resetting all blocks' AStarParents after completing a path, there
could be an infinite loop in certain functions. Also is this going to cause problems if you switch between
different paths between ticks?

* Fixed bug where colliding blocks/modules were picked up diagonally as well as adjacently.

* After several attempts at rewriting Tower::FindNewParent(), added Tower::FindNewParentAStar() (temp name),
which uses an AStarComponent to find a new parent. Funny how I never realized until now that a block-only
A* search was exactly what I needed.

* Added PR_Goal_ConnectedToRoot A* rule, the first goal rule. While paths with all the other rulesets
fail if they can't get to the passed in Finish location, this rule immediately returns a successful
path upon the first block it finds connected to the root.

* Fixed bug where blocks would fall through blocks under them that weren't connected to the root. Now they
parent with them.

* ToDo:

- Resolve all bugs with the block system. At the very least, sometimes some blocks fall into other blocks rarely. Fix it.

- Finish up AStarComponent. Fix probable AStarParent bug.

- Lose the Tower prefix.

- Loading support for new JSON system.